Move find_model_with_retry and use in mpl-core to fix loop
