+++ program_transformers/src/lib.rs
@@ -16,11 +16,15 @@ use {
         },
     },
     futures::future::BoxFuture,
-    sea_orm::{DatabaseConnection, SqlxPostgresConnector},
+    sea_orm::{
+        entity::EntityTrait, query::Select, ConnectionTrait, DatabaseConnection, DbErr,
+        SqlxPostgresConnector, TransactionTrait,
+    },
     solana_sdk::{instruction::CompiledInstruction, pubkey::Pubkey, signature::Signature},
     solana_transaction_status::InnerInstructions,
     sqlx::PgPool,
     std::collections::{HashMap, HashSet, VecDeque},
+    tokio::time::{sleep, Duration},
     tracing::{debug, error, info},
 };
 
@@ -245,5 +249,38 @@ impl ProgramTransformer {
             }?;
         }
         Ok(())
+    }
+}
+
+pub async fn find_model_with_retry<T: ConnectionTrait + TransactionTrait, K: EntityTrait>(
+    conn: &T,
+    model_name: &str,
+    select: &Select<K>,
+    retry_intervals: &[u64],
+) -> Result<Option<K::Model>, DbErr> {
+    let mut retries = 0;
+    let metric_name = format!("{}_found", model_name);
+
+    for interval in retry_intervals {
+        let interval_duration = Duration::from_millis(*interval);
+        sleep(interval_duration).await;
+
+        let model = select.clone().one(conn).await?;
+        if let Some(m) = model {
+            record_metric(&metric_name, true, retries);
+            return Ok(Some(m));
+        }
+        retries += 1;
+    }
+
+    record_metric(&metric_name, false, retries - 1);
+    Ok(None)
+}
+
+fn record_metric(metric_name: &str, success: bool, retries: u32) {
+    let retry_count = &retries.to_string();
+    let success = if success { "true" } else { "false" };
+    if cadence_macros::is_global_default_set() {
+        cadence_macros::statsd_count!(metric_name, 1, "success" => success, "retry_count" => retry_count);
     }
 }
