+++ program_transformers/src/token_metadata/v1_asset.rs
@@ -6,7 +6,7 @@ use {
             AssetMintAccountColumns, AssetTokenAccountColumns,
         },
         error::{ProgramTransformerError, ProgramTransformerResult},
-        DownloadMetadataInfo,
+        find_model_with_retry, DownloadMetadataInfo,
     },
     blockbuster::token_metadata::{
         accounts::{MasterEdition, Metadata},
@@ -26,12 +26,11 @@ use {
     },
     sea_orm::{
         entity::{ActiveValue, ColumnTrait, EntityTrait},
-        query::{JsonValue, Order, QueryFilter, QueryOrder, QueryTrait, Select},
+        query::{JsonValue, Order, QueryFilter, QueryOrder, QueryTrait},
         sea_query::query::OnConflict,
         ConnectionTrait, DbBackend, DbErr, TransactionTrait,
     },
     solana_sdk::{pubkey, pubkey::Pubkey},
-    tokio::time::{sleep, Duration},
     tracing::warn,
 };
 
@@ -409,37 +408,4 @@ pub async fn save_v1_asset<T: ConnectionTrait + TransactionTrait>(
     }
 
     Ok(Some(DownloadMetadataInfo::new(mint_pubkey_vec, uri)))
-}
-
-async fn find_model_with_retry<T: ConnectionTrait + TransactionTrait, K: EntityTrait>(
-    conn: &T,
-    model_name: &str,
-    select: &Select<K>,
-    retry_intervals: &[u64],
-) -> Result<Option<K::Model>, DbErr> {
-    let mut retries = 0;
-    let metric_name = format!("{}_found", model_name);
-
-    for interval in retry_intervals {
-        let interval_duration = Duration::from_millis(*interval);
-        sleep(interval_duration).await;
-
-        let model = select.clone().one(conn).await?;
-        if let Some(m) = model {
-            record_metric(&metric_name, true, retries);
-            return Ok(Some(m));
-        }
-        retries += 1;
-    }
-
-    record_metric(&metric_name, false, retries - 1);
-    Ok(None)
-}
-
-fn record_metric(metric_name: &str, success: bool, retries: u32) {
-    let retry_count = &retries.to_string();
-    let success = if success { "true" } else { "false" };
-    if cadence_macros::is_global_default_set() {
-        cadence_macros::statsd_count!(metric_name, 1, "success" => success, "retry_count" => retry_count);
-    }
 }
